requirements.txt
flask
flask-cors
supabase
imagekitio
argon2-cffi
requests
gunicorn
wheel
boxsdk
python-dotenv
dropbox